+++ Config/AutoZeroMode.cs
@@ -1,0 +1,9 @@
+namespace ScopeRangefinder
+{
+    public enum AutoZeroMode
+    {
+        Hotkey,
+
+        Continuous
+    }
+}