+++ Config/ConfigurationManagerAttributes.cs
@@ -1,0 +1,8 @@
+namespace ScopeRangefinder
+{
+    internal sealed class ConfigurationManagerAttributes
+    {
+        public string DispName;
+        public int? Order;
+    }
+}